Refactor evaluation and linking scripts
There exists now a class LinkingSystem that combines the three
components of our linker (link-linker, linker, coref-linker).
The script link_benchmark_entities.py links benchmark input and writes
the result together with ground truth labels to a file.
A script link_entities.py links entities in raw or article-input and
writes the result to a file.
A script evaluate_linked_entities.py takes the output of
link_benchmark_entities.py, prints an evaluation and writes the result
evaluation cases to a file.