Show types as they are counted in the evaluation
In the evaluation, only the type of the parent GT entity counts for TP
und FN. Now, when selecting a type in the table, those mentions are
highlighted that count towards this type (e.g. not a child label that
has this type but does not count as it because its parent has a
different type).