First version of a subgraph visualizer using D3
Web app to visualize the subgraph induced by a given entity via
wdt:P31/wdt:P279 on our wikidata-types entity (which has a reduced set
of types and various corrections compared to Wikidata).

The basic functionality is there for a fixed entity. Still needs some
work, in particular:

1. Nicer web page (currently very rudimentary)
2. nicer graph (no effort spent on proper styling so far)
3. Select the entity via autocompletion (right now, the URL is
hard-coded for a fixe entity)