Improve graph builder + pick entity via autocompletion
1. The layout of the graph builder is now much improved. In particular:

1.1 Nice arrow heads (these have to be drawn by hand)
1.2 Variable node sizes (grows with logarithm of number of sitelinks)
1.3 Arrow heads end before node circle (tricky)
1.4 Edges in different colors, depending on predicate
1.5 Nodes in different colors, depending on type (source or not)
1.6 Tooltip for nodes and the (abbreviated) labels
1.7 Labels can be clicked -> Wikidata page for entity
1.8 Zoom and pan (works, but strange things happen when beyond viewport)

2. The entity can now be picked via autocompletion:

2.1 QLever is used as backend, it's pretty fast. I first tried
Wikimedia's wbsearchentities service, but got a CORS-related error
message, which I could not resolve. Strangely enough,
https://angryloki.github.io/wikidata-graph-builder uses wbsearchentities
and it works there.
2.2 I copied the code for the AC dropdown query from
https://www.w3schools.com/howto/howto_js_autocomplete.asp . It was given
for a fixed array of possible completions. Extending that to completions
obtained via fetch was relatively easy.

3. Improved code + working with async and await for better code
structure.

The biggest remaining TODO is the strange behaviour of zoom and pan,
when parts of the graph are outside the viewport. I tried clipping, but
that did not work as expected (the clip rectangle moved with the pan,
which it should not). I would like arbitrary zoom and pan (but clipped)
like angryloki does it.