Use mappings from QID to property instead of WikidataEntity properties
That means self.entities in EntityDatabase is now only a set of relevant
entities (necessary in some rare and rather outdated cases for the
contains_entity method).
It also means, WikidataEntity contains now only ID, name and type
information, which are needed for storing such thatthese properties can
be read by the webapp.