Store true basekey, key and subkey in data instead of trf'ing to class
Before, the cell's class was used to determine its header, tooltip etc.
But the class was not always equal to the actual json key, since the
json key might contain characters that are forbidden in html class
names. This can be a problem when custom ontology types are used.
Now the basekey, key and subkey of the corresponding json entry are
stored in the cell's data object.