Fix multiprocessing by cleaning up stuck processes every x articles
Also use imap_unordered() instead of imap(). The problem before with the
ever increasing RAM usage at some point was not (as suspected) a memory
leak as they happen with multiprocessing, the problem was that one
process got stuck and the other processes could not write their results
until the stuck process wrote its results (which never happened). So the
other processes kept reading articles into memory but never released
them.
Now i use imap_unordered(). If a process gets stuck, the other processes
will not wait for it to write its results, but just write theirs when
they are done. This also means, that the output is not ordered anymore.
After x=50,000 articles all processes are stopped and new ones are
created. This way, a small fraction of articles can get lost, but we
don't have the linking process will actually terminate (unless one is
very unlucky and all processes get stuck at some point).