Use link frequencies to get most popular candidate instead of sitelinks
Before, when using sitelinks to determine the most popular candidate, it
was enough if a more popular entity was linked once with the alias in
Wikipedia. E.g. France was a rare error, because it was at least once
falsely linked to Turkey in Wikipedia.

This affects rare and metonym errors.