Adjust code to train a spaCy model to spaCy version 3.0
the old code was not running anymore for the new version.
There must be a problem with the new code as well however, since the
training loss is always 0. I was unable to figure out why until now.
I followed the instructions in the official spaCy entity linking
tutorial here:
https://github.com/explosion/projects/blob/v3/tutorials/nel_emerson/notebooks/notebook_video.ipynb

Relevant to #14